+++ CHANGELOG.rst
@@ -13,6 +13,8 @@ Added
 
 - **Samplesheets**
     - ``save_cache`` arg in ``SampleSheetTableBuilder.get_study_tables()`` (#2190)
+- **Taskflowbackend**
+    - Percentages in zone progress counters (#2189)
 
 Changed
 -------